requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==2.2.9
djangorestframework==3.11.0
idna==2.9
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pkg-resources==0.0.0
pytz==2019.3
requests==2.23.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
